+++ ui/src/app/components/runtime-configuration-panel/customize-panel-footer.spec.tsx
@@ -5,7 +5,8 @@ import * as React from 'react';
 import { DiskType, RuntimeStatus } from 'generated/fetch';
 
 import { render, screen, waitFor } from '@testing-library/react';
-import { PanelContent, toAnalysisConfig } from 'app/utils/runtime-utils';
+import { toAnalysisConfig } from 'app/utils/analysis-config';
+import { PanelContent } from 'app/utils/runtime-utils';
 import { serverConfigStore } from 'app/utils/stores';
 
 import defaultServerConfig from 'testing/default-server-config';
