+++ ui/src/app/components/runtime-initializer-modal.tsx
@@ -5,8 +5,8 @@ import { Runtime, RuntimeConfigurationType } from 'generated/fetch';
 
 import colors, { colorWithWhiteness } from 'app/styles/colors';
 import { reactStyles } from 'app/utils';
+import { toAnalysisConfig } from 'app/utils/analysis-config';
 import { setSidebarActiveIconStore } from 'app/utils/navigation';
-import { toAnalysisConfig } from 'app/utils/runtime-utils';
 import { runtimeDiskStore, useStore } from 'app/utils/stores';
 
 import { Button, Clickable } from './buttons';
