+++ ui/src/app/utils/machines.ts
@@ -4,8 +4,9 @@ import { Disk, DiskType } from 'generated/fetch';
 
 import { DEFAULT, switchCase } from '@terra-ui-packages/core-utils';
 
+import { AnalysisConfig } from './analysis-config';
 import { formatUsd } from './numbers';
-import { AnalysisConfig, DiskConfig } from './runtime-utils';
+import { DiskConfig } from './runtime-utils';
 
 // Copied from https://github.com/DataBiosphere/terra-ui/blob/219b063b07d56499ccc38013fd88f4f0b88f8cd6/src/data/machines.js
